+++ src/main/java/org/sambit/kids/math/equations/Equation.java
@@ -1,4 +1,4 @@
-package org.sambit.kids.math;
+package org.sambit.kids.math.equations;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
