+++ src/main/java/org/sambit/kids/math/equations/EquationsGenerator.java
@@ -1,4 +1,4 @@
-package org.sambit.kids.math;
+package org.sambit.kids.math.equations;
 
 import java.util.List;
